# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Open API - E Commerce",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "darius@matulionis.lt"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
...
        "/api/product": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Show All product",
                "description": "Returns a all Product",
                "operationId": "be77b9897465d0527c2d85bc9cef84f9",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "title": "data",
                                    "type": "array",
                                    "items": {
                                        "title": "data[0]",
                                        "properties": {
                                            "id": {
                                                "type": "string",
                                                "example": "1"
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Samsung S24"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "samsung memiliki ....."
                                            },
                                            "price": {
                                                "type": "integer",
                                                "example": "34000000"
                                            },
                                            "stock": {
                                                "type": "integer",
                                                "example": "8"
                                            },
                                            "rate": {
                                                "type": "integer",
                                                "example": "5"
                                            },
                                            "category_id": {
                                                "type": "integer",
                                                "example": "1"
                                            },
                                            "category": {
                                                "type": "string",
                                                "example": "Samsung"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "example": [
                                    {
                                        "id": "1",
                                        "name": "Samsung S24",
                                        "description": "Samsung S24 memiliki...",
                                        "price": "3000000",
                                        "stock": "10",
                                        "rate": "4",
                                        "category_id": "1",
                                        "category": "Samsung"
                                    },
                                    {
                                        "id": "2",
                                        "name": "Samsung S23",
                                        "description": "Samsung S23 memiliki...",
                                        "price": "2800000",
                                        "stock": "10",
                                        "rate": "4",
                                        "category_id": "1",
                                        "category": "Samsung"
                                    }
                                ]
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Product not Found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Create new product",
                "description": "Create a new product",
                "operationId": "a2a748b2864d35d1e50976b938ba8cd9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Samsung S24"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "samsung memiliki ....."
                                    },
                                    "price": {
                                        "type": "integer",
                                        "example": "34000000"
                                    },
                                    "stock": {
                                        "type": "integer",
                                        "example": "8"
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "example": "1"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Samsung S24"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "samsung memiliki ....."
                                        },
                                        "price": {
                                            "type": "integer",
                                            "example": "34000000"
                                        },
                                        "stock": {
                                            "type": "integer",
                                            "example": "8"
                                        },
                                        "rate": {
                                            "type": "integer",
                                            "example": "null"
                                        },
                                        "category_id": {
                                            "type": "integer",
                                            "example": "1"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/api/product/{Id}": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Find Product by ID",
                "description": "Returns a single product",
                "operationId": "3a9c6bb5e73d3518eac56cee12301a56",
                "parameters": [
                    {
                        "name": "Id",
                        "in": "path",
                        "description": "Id Product",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "example": "1"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Samsung S24"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "samsung memiliki ....."
                                        },
                                        "price": {
                                            "type": "integer",
                                            "example": "34000000"
                                        },
                                        "stock": {
                                            "type": "integer",
                                            "example": "8"
                                        },
                                        "rate": {
                                            "type": "integer",
                                            "example": "5"
                                        },
                                        "category_id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "category": {
                                            "type": "string",
                                            "example": "Samsung"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid ID supplier"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Product"
                ],
                "summary": "Delete Product by ID",
                "description": "Delete a single product",
                "operationId": "64671fca2ad1591c7ab3e0ee093ac402",
                "parameters": [
                    {
                        "name": "Id",
                        "in": "path",
                        "description": "Id Product",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid ID supplier"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            }
        },
...
    },
...
}